updating to use nmdc_api_utilities, removing api classes
