+++ requirements.txt
@@ -7,3 +7,4 @@ tqdm~=4.66.5
 pandas~=2.2.3
 python-dotenv==1.0.1
 typing-inspect~=0.9.0
+nmdc_api_utilities
